Fix group models. AcademicYear data type short -> int

+++ ELearning.Application/Groups/Queries/GetGroupsList/GroupLookupModel.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public short AcademicYear { get; set; }
+        public int AcademicYear { get; set; }
         public int SubjectId { get; set; }
         public string SubjectName { get; set; }
         public string SubjectAbreviation { get; set; }
